+++ pkg/model/database/database.go
@@ -80,6 +80,7 @@ var ProjectFields = struct {
 	AvatarURL       string
 	TokenID         string
 	Token           string
+	ProviderID      string
 	Provider        string
 	BuildDefinition string
 	Branches        string
@@ -92,6 +93,7 @@ var ProjectFields = struct {
 	AvatarURL:       "AvatarURL",
 	TokenID:         "TokenID",
 	Token:           "Token",
+	ProviderID:      "ProviderID",
 	Provider:        "Provider",
 	BuildDefinition: "BuildDefinition",
 	Branches:        "Branches",
@@ -102,9 +104,19 @@ var ProjectFields = struct {
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
 var ProjectColumns = struct {
-	TokenID string
+	ProjectID   string
+	Name        string
+	GroupName   string
+	Description string
+	TokenID     string
+	GitURL      string
 }{
-	TokenID: "token_id",
+	ProjectID:   "project_id",
+	Name:        "name",
+	GroupName:   "group_name",
+	Description: "description",
+	TokenID:     "token_id",
+	GitURL:      "git_url",
 }
 
 // Project holds data about an imported project. A lot of the data is expected
@@ -116,9 +128,9 @@ type Project struct {
 	GroupName       string    `gorm:"size:500;not null;default:''"`
 	Description     string    `gorm:"size:500;not null;default:''"`
 	AvatarURL       string    `gorm:"size:500;not null;default:''"`
-	TokenID         uint      `gorm:"nullable;default:NULL;index:project_idx_token_id"`
+	TokenID         *uint     `gorm:"nullable;default:NULL;index:project_idx_token_id"`
 	Token           *Token    `gorm:"foreignKey:TokenID;constraint:OnUpdate:RESTRICT,OnDelete:RESTRICT"`
-	ProviderID      uint      `gorm:"nullable;default:NULL;index:project_idx_provider_id"`
+	ProviderID      *uint     `gorm:"nullable;default:NULL;index:project_idx_provider_id"`
 	Provider        *Provider `gorm:"foreignKey:ProviderID;constraint:OnUpdate:RESTRICT,OnDelete:RESTRICT"`
 	BuildDefinition string    `gorm:"not null;default:''"`
 	Branches        []Branch  `gorm:"foreignKey:ProjectID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE"`